+++ packages/react/CHANGELOG.md
@@ -8,7 +8,11 @@ description: All notable changes will be documented in this file.
 
 ### Fixed
 
-- **Steps**: Fixed issue where steps context was not exported.
+- **Steps**: Fixed issue where steps context was not exported
+
+## Added
+
+- **Checkbox**: Added `invalid` prop to `Checkbox.Group`
 
 ## [3.9.0] - 2024-08-22